Move package verification logic to CommonRuntimeImageTests
